Added support for resizable terminals
This patch contains modificatio to two source files, acceptch.c
and screen.c.

In acceptch.c, there is a if-block on a global variable SIZE_CH,
which is by default set to FALSE and only set to TRUE when
handling the terminal resize signal.  In previous versions,
this block provided states initialization but no other functions,
so resizable terminals was not supported.  This patch fix these
states, so that resize handler routine "checkScreenSize" in
screen.c can update the states correctly.

In screen.c, this patch re-calculates key variables like
"currentLine", remain the position after resizing a working
terminal.

Known bugs:
* long press up/down keys until the top/bottom, chances are that
  the underscore will still show on the 2nd/ the 2nd-last line.
* After multiple terminal resizing, the first input key catched by
  wgetch() in acceptch.c will somehow be transformed into a
  sequence of code, which will be interpreted as some ASCII
  sequence, and cause undesirable modifications.
  This is confirmed using gnome-terminal and lxterminal in Linux.