fix format truncation error with GCC-7
GCC-7 introduced new warnings and errors. Among them is a new warning
for possible truncations in the output of snprintf. Since we are only
interested in the return value of snprintf and do not use the string
written by it we can also replace the buffer with a NULL pointer.
This makes it explicit that we do not want to write a string and
silences the GCC-7 error.

See also the examples in
http://en.cppreference.com/w/c/io/fprintf